tests: Split array testsets
