+++ test/array/core.jl
@@ -1,9 +1,3 @@
-using LinearAlgebra, SparseArrays, Random, SharedArrays
-import Dagger: DArray, chunks, domainchunks, treereduce_nd
-import Distributed: myid, procs
-import Statistics: mean, var, std
-import OnlineStats
-
 @testset "treereduce_nd" begin
     xs = rand(1:10, 8,8,8)
     concats = [(x...)->cat(x..., dims=n) for n in 1:3]
@@ -80,52 +74,6 @@ end
     end
 end
 
-function test_mapreduce(f, init_func; no_init=true, zero_init=zero,
-                        types=(Int32, Int64, Float32, Float64),
-                        cmp=isapprox)
-    @testset "$T" for T in types
-        X = init_func(Blocks(10, 10), T, 100, 100)
-        inits = ()
-        if no_init
-            inits = (inits..., nothing)
-        end
-        if zero_init !== nothing
-            inits = (inits..., zero_init(T))
-        end
-        @testset "dims=$dims" for dims in (Colon(), 1, 2, (1,), (2,))
-            @testset "init=$init" for init in inits
-                if init === nothing
-                    if dims == Colon()
-                        @test cmp(f(X; dims), f(collect(X); dims))
-                    else
-                        @test cmp(collect(f(X; dims)), f(collect(X); dims))
-                    end
-                else
-                    if dims == Colon()
-                        @test cmp(f(X; dims, init), f(collect(X); dims, init))
-                    else
-                        @test cmp(collect(f(X; dims, init)), f(collect(X); dims, init))
-                    end
-                end
-            end
-        end
-    end
-end
-
-# Base
-@testset "reduce" test_mapreduce((X; dims, init=Base._InitialValue())->reduce(+, X; dims, init), ones)
-@testset "mapreduce" test_mapreduce((X; dims, init=Base._InitialValue())->mapreduce(x->x+1, +, X; dims, init), ones)
-@testset "sum" test_mapreduce(sum, ones)
-@testset "prod" test_mapreduce(prod, rand)
-@testset "minimum" test_mapreduce(minimum, rand)
-@testset "maximum" test_mapreduce(maximum, rand)
-@testset "extrema" test_mapreduce(extrema, rand; cmp=Base.:(==), zero_init=T->(zero(T), zero(T)))
-
-# Statistics
-@testset "mean" test_mapreduce(mean, rand; zero_init=nothing, types=(Float32, Float64))
-@testset "var" test_mapreduce(var, rand; zero_init=nothing, types=(Float32, Float64))
-@testset "std" test_mapreduce(std, rand; zero_init=nothing, types=(Float32, Float64))
-
 @testset "broadcast" begin
     X1 = rand(Blocks(10), 100)
     X2 = X1 .* 3.4
@@ -138,7 +86,7 @@ end
 
 @testset "distributing an array" begin
     function test_dist(X)
-        X1 = distribute(X, Blocks(10, 20))
+        X1 = Distribute(Blocks(10, 20), X)
         Xc = fetch(X1)
         @test Xc isa DArray{eltype(X),ndims(X)}
         @test Xc == X
@@ -147,7 +95,7 @@ end
         @test map(x->size(x) == (10, 20), domainchunks(Xc)) |> all
     end
     x = [1 2; 3 4]
-    @test distribute(x, Blocks(1,1)) == x
+    @test Distribute(Blocks(1,1), x) == x
     test_dist(rand(100, 100))
     test_dist(sprand(100, 100, 0.1))
 
@@ -171,39 +119,10 @@ end
     test_transpose(sprand(100, 120, 0.1))
 end
 
-@testset "matrix-matrix multiply" begin
-    function test_mul(X)
-        tol = 1e-12
-        X1 = distribute(X, Blocks(10, 20))
-        @test_throws DimensionMismatch X1*X1
-        X2 = X1'*X1
-        X3 = X1*X1'
-        @test norm(collect(X2) - X'X) < tol
-        @test norm(collect(X3) - X*X') < tol
-        @test chunks(X2) |> size == (2, 2)
-        @test chunks(X3) |> size == (4, 4)
-        @test map(x->size(x) == (20, 20), domainchunks(X2)) |> all
-        @test map(x->size(x) == (10, 10), domainchunks(X3)) |> all
-    end
-    test_mul(rand(40, 40))
-
-    x = rand(10,10)
-    X = distribute(x, Blocks(3,3))
-    y = rand(10)
-    @test norm(collect(X*y) - x*y) < 1e-13
-end
-
-@testset "matrix powers" begin
-    x = rand(Blocks(4,4), 16, 16)
-    @test collect(x^1) == collect(x)
-    @test collect(x^2) == collect(x*x)
-    @test collect(x^3) == collect(x*x*x)
-end
-
 @testset "concat" begin
     m = rand(75,75)
-    x = distribute(m, Blocks(10,20))
-    y = distribute(m, Blocks(10,10))
+    x = Distribute(Blocks(10,20), m)
+    y = Distribute(Blocks(10,10), m)
     @test hcat(m,m) == collect(hcat(x,x)) == collect(hcat(x,y))
     @test vcat(m,m) == collect(vcat(x,x))
     @test_throws DimensionMismatch vcat(x,y)
@@ -211,7 +130,7 @@ end
 
 @testset "scale" begin
     x = rand(10,10)
-    X = distribute(x, Blocks(3,3))
+    X = Distribute(Blocks(3,3), x)
     y = rand(10)
 
     @test Diagonal(y)*x == collect(Diagonal(y)*X)
@@ -219,7 +138,7 @@ end
 
 @testset "Getindex" begin
     function test_getindex(x)
-        X = distribute(x, Blocks(3,3))
+        X = Distribute(Blocks(3,3), x)
         @test collect(X[3:8, 2:7]) == x[3:8, 2:7]
         ragged_idx = [1,2,9,7,6,2,4,5]
         @test collect(X[ragged_idx, 2:7]) == x[ragged_idx, 2:7]
@@ -248,7 +167,7 @@ end
 
 
 @testset "cleanup" begin
-    X = distribute(rand(10,10), Blocks(10,10))
+    X = Distribute(Blocks(10,10), rand(10,10))
     @test collect(sin.(X)) == collect(sin.(X))
 end
 
@@ -269,7 +188,7 @@ end
     x=rand(10,10)
     y=copy(x)
     y[3:8, 2:7] .= 1.0
-    X = distribute(x, Blocks(3,3))
+    X = Distribute(Blocks(3,3), x)
     @test collect(setindex(X,1.0, 3:8, 2:7)) == y
     @test collect(X) == x
 end
@@ -292,7 +211,7 @@ end
     @test collect(sort(y)) == x
 
     x = ones(10)
-    y = distribute(x, Blocks(3))
+    y = Distribute(Blocks(3), x)
     @test_broken map(x->length(collect(x)), sort(y).chunks) == [3,3,3,1]
 end
